Moved packages and documentation
