+++ app/src/main/java/com/apptec/registrateapp/mainactivity/fnotification/ui/NotificationView.java
@@ -1,16 +1,18 @@
-package com.apptec.registrateapp.mainactivity.fnotification;
+package com.apptec.registrateapp.mainactivity.fnotification.ui;
 
 public interface NotificationView {
     /*
-    * Notification View interface
-    * */
-
+     * Notification View interface
+     * */
 
 
     void showNotNewNotificationsMessage();
 
     void showAssistanceProgressDialog(String string);
+
     void hideAssistanceProgressDialog();
+
     void showConnectionErrorMessage();
+
     void showDialog(String title, String message);
 }
