+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -15,7 +15,7 @@
     </fragment>
     <fragment
         android:id="@+id/notificationsFragment"
-        android:name="com.apptec.registrateapp.mainactivity.fnotification.NotificationFragment"
+        android:name="com.apptec.registrateapp.mainactivity.fnotification.ui.NotificationFragment"
         android:label="NotificationsFragment">
         <action
             android:id="@+id/action_notificationsFragment_to_permissionFragment"
@@ -23,7 +23,7 @@
     </fragment>
     <fragment
         android:id="@+id/permissionFragment"
-        android:name="com.apptec.registrateapp.mainactivity.fpermission.PermissionFragment"
+        android:name="com.apptec.registrateapp.mainactivity.fpermission.ui.PermissionFragment"
         android:label="PermissionFragment">
         <action
             android:id="@+id/action_permissionFragment_to_dialogDevice"
